Extract testing tools, split model tests in diff files.
